+++ packages/firestore/test/integration/api/query.test.ts
@@ -62,6 +62,7 @@ import {
   toChangesArray,
   toDataArray,
   toIds,
+  PERSISTENCE_MODE_UNSPECIFIED,
   withEmptyTestCollection,
   withRetry,
   withTestCollection,
@@ -2110,16 +2111,18 @@ apiDescribe('Queries', persistence => {
               snapshot => snapshot.ref
             );
 
-            // Delete 50 of the 100 documents. Use a WriteBatch, rather than
-            // deleteDoc(), to avoid affecting the local cache.
+            // Delete 50 of the 100 documents. Use a different Firestore
+            // instance to avoid affecting the local cache.
             const deletedDocumentIds = new Set<string>();
-            const writeBatchForDocumentDeletes = writeBatch(db);
-            for (let i = 0; i < createdDocuments.length; i += 2) {
-              const documentToDelete = createdDocuments[i];
-              writeBatchForDocumentDeletes.delete(documentToDelete);
-              deletedDocumentIds.add(documentToDelete.id);
-            }
-            await writeBatchForDocumentDeletes.commit();
+            await withTestDb(PERSISTENCE_MODE_UNSPECIFIED, async db2 => {
+              const batch = writeBatch(db2);
+              for (let i = 0; i < createdDocuments.length; i += 2) {
+                const documentToDelete = doc(db2, createdDocuments[i].path);
+                batch.delete(documentToDelete);
+                deletedDocumentIds.add(documentToDelete.id);
+              }
+              await batch.commit();
+            });
 
             // Wait for 10 seconds, during which Watch will stop tracking the
             // query and will send an existence filter rather than "delete"
@@ -2262,11 +2265,12 @@ apiDescribe('Queries', persistence => {
         );
 
         // Delete one of the documents so that the next call to getDocs() will
-        // experience an existence filter mismatch. Use a WriteBatch, rather
-        // than deleteDoc(), to avoid affecting the local cache.
-        const writeBatchForDocumentDeletes = writeBatch(db);
-        writeBatchForDocumentDeletes.delete(doc(coll, 'DocumentToDelete'));
-        await writeBatchForDocumentDeletes.commit();
+        // experience an existence filter mismatch. Use a different Firestore
+        // instance to avoid affecting the local cache.
+        const documentToDelete = doc(coll, 'DocumentToDelete');
+        await withTestDb(PERSISTENCE_MODE_UNSPECIFIED, async db2 => {
+          await deleteDoc(doc(db2, documentToDelete.path));
+        });
 
         // Wait for 10 seconds, during which Watch will stop tracking the query
         // and will send an existence filter rather than "delete" events when
@@ -2282,7 +2286,7 @@ apiDescribe('Queries', persistence => {
           documentSnapshot => documentSnapshot.id
         );
         const testDocIdsMinusDeletedDocId = testDocIds.filter(
-          documentId => documentId !== 'DocumentToDelete'
+          documentId => documentId !== documentToDelete.id
         );
         expect(snapshot2DocumentIds, 'snapshot2DocumentIds').to.have.members(
           testDocIdsMinusDeletedDocId
@@ -2313,20 +2317,17 @@ apiDescribe('Queries', persistence => {
         // is if there is a false positive when testing for 'DocumentToDelete'
         // in the bloom filter. So verify that the bloom filter application is
         // successful, unless there was a false positive.
-        const isFalsePositive = bloomFilter.mightContain!(
-          `${coll.path}/DocumentToDelete`
-        );
+        const isFalsePositive = bloomFilter.mightContain(documentToDelete);
         expect(bloomFilter.applied, 'bloomFilter.applied').to.equal(
           !isFalsePositive
         );
 
         // Verify that the bloom filter contains the document paths with complex
         // Unicode characters.
-        for (const testDocId of testDocIdsMinusDeletedDocId) {
-          const testDocPath = `${coll.path}/${testDocId}`;
+        for (const testDoc of snapshot2.docs.map(snapshot => snapshot.ref)) {
           expect(
-            bloomFilter.mightContain!(testDocPath),
-            `bloomFilter.mightContain('${testDocPath}')`
+            bloomFilter.mightContain(testDoc),
+            `bloomFilter.mightContain('${testDoc.path}')`
           ).to.be.true;
         }
       });